pyproject.toml
[project]
name = "atrax-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.6.0",
    "django>=5.2.9",
    "django-celery-beat>=2.8.1",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "psycopg2-binary>=2.9.11",
    "redis>=7.1.0",
    "requests>=2.32.5",
]
